Refactor imports: update tech model imports to type imports for improved clarity
